购物车，收货地址完成

+++ graduation-shop/src/main/java/com/jd/graduation/Impl/CartServiceImpl.java
@@ -2,6 +2,7 @@ package com.jd.graduation.Impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.jd.graduation.DO.CartDO;
+import com.jd.graduation.DTO.CartChangeAmountDTO;
 import com.jd.graduation.VO.BookVO;
 import com.jd.graduation.service.CartService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,14 +27,21 @@ public class CartServiceImpl extends CartService {
         cartDO.setUid(uid);
         cartDO.setBid(bid);
         cartDO.setPrice(bookVO.getPrice());
+        cartDO.setAmount(1);
         cartDO.setCreateTime(new Date());
 
         baseMapper.insert(cartDO);
         return null;
     }
 
-    public void deleteFromCart(Integer id) {
+    public String delete(Integer id, Integer uid) {
+        CartDO cartDO = baseMapper.selectById(id);
+        if (cartDO == null || !cartDO.getUid().equals(uid)){
+            return "不是你的购物车";
+        }
+
         baseMapper.deleteById(id);
+        return null;
     }
 
     public List<CartDO> getInIds(List<Integer> ids, Integer uid) {
@@ -44,5 +52,16 @@ public class CartServiceImpl extends CartService {
         QueryWrapper<CartDO> wrapper = new QueryWrapper<>();
         wrapper.eq("uid", uid);
         return baseMapper.selectList(wrapper);
+    }
+
+    public String changeAmount(CartChangeAmountDTO dto, Integer uid) {
+        CartDO cartDO = baseMapper.selectById(dto.getId());
+        if (cartDO == null || !cartDO.getUid().equals(uid)){
+            return "不是你的购物车";
+        }
+
+        cartDO.setAmount(dto.getAmount());
+        baseMapper.updateById(cartDO);
+        return null;
     }
 }
